code refactor, added .gitignore
