+++ Lab1_SortingAlgorithms/sort.py
@@ -1,9 +1,16 @@
 import argparse
-import sorter
+from Lab1_SortingAlgorithms import sorter
+
 ALLOWED_ALGORITHMS = ["quick", "selection"]
+
+
 parser = argparse.ArgumentParser(description="sort integer array")
+
 parser.add_argument("list", help="list to be sorted", type=float, nargs='+')
 parser.add_argument("-m", "--method", help="sorting algorithm to be used", choices=ALLOWED_ALGORITHMS)
+
 args = parser.parse_args()
+
 sorter.sort(args.list, args.method)
+
 print(args.list)
